Section about your own project presentation
According to my experience of interview, such presentation is must have. That's the one of keystones in successful interview, so it must be prepared properly. 
Also I'll be grateful if you can share your own experience of completing project on the start of your career (where to get them and what to focus on etc.).
Thanks for great guidance!